Add language with support for variable binding
The language defined in 'bound.rkt' extends the base language with
support for variable bindings and references. This module then defines
both a reduction relation on this extended language as an extension of
the base/eval relation, performing beta-reduction on variable bindings,
as well as a translation from the bound language to the base language,
demonstrating the binding form need not be included in the foundational
calculus in order to support variables.